requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
motor==3.3.2
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-dotenv==1.0.0
openai==1.3.7
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
gspread==5.12.0
celery==5.3.4
redis==5.0.1
twilio==8.11.0
httpx==0.25.2
websockets==12.0
cryptography>=41.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
requests==2.31.0